feat(scalars): improve types & detach booleans
