+++ packages/design-system/src/scalars/components/boolean-field/index.ts
@@ -1,1 +1,1 @@
-export * from "./boolean-field.js";
+export { BooleanField } from "./boolean-field.js";
