+++ packages/design-system/src/scalars/components/fragments/checkbox-field/checkbox-base.test.tsx
@@ -1,16 +1,16 @@
-import { render, screen, fireEvent } from "@testing-library/react";
-import { describe, it, expect, vi } from "vitest";
-import { Checkbox } from "./checkbox.js";
+import { fireEvent, render, screen } from "@testing-library/react";
+import { describe, expect, it, vi } from "vitest";
+import { CheckboxBase } from "./checkbox-base.js";
 
 describe("Checkbox", () => {
   it("should render a checkbox", () => {
-    render(<Checkbox />);
+    render(<CheckboxBase />);
     expect(screen.getByRole("checkbox")).toBeInTheDocument();
   });
 
   it("should handle checked state", () => {
     const handleCheckedChange = vi.fn();
-    render(<Checkbox onCheckedChange={handleCheckedChange} />);
+    render(<CheckboxBase onCheckedChange={handleCheckedChange} />);
 
     const checkbox = screen.getByRole("checkbox");
     fireEvent.click(checkbox);
@@ -18,7 +18,7 @@ describe("Checkbox", () => {
   });
 
   it("should be disabled when disabled prop is true", () => {
-    render(<Checkbox disabled />);
+    render(<CheckboxBase disabled />);
     expect(screen.getByRole("checkbox")).toBeDisabled();
   });
 });
