+++ packages/design-system/src/scalars/components/fragments/checkbox-field/index.ts
@@ -1,2 +1,2 @@
-export * from "./checkbox.js";
-export * from "./checkbox-field.js";
+export { CheckboxField } from "./checkbox-field.js";
+export { Checkbox } from "./checkbox.js";
