+++ packages/design-system/src/scalars/docs/examples/enum-field-example/enum-field-example.tsx
@@ -1,57 +1,53 @@
 import { Button } from "#powerhouse";
-import { EnumField, Form } from "#scalars";
+import { BooleanField, Checkbox, Form, Toggle } from "#scalars";
+import { useState } from "react";
 
 const EnumFieldExample = () => {
   const onSubmit = async (data: any) => {
     // simulate a network request
-    await new Promise((resolve) => setTimeout(resolve, 1000));
     alert(JSON.stringify(data, null, 2));
   };
 
-  const genderOptions = [
-    { label: "Male", value: "male" },
-    { label: "Female", value: "female" },
-    { label: "Other", value: "other" },
-  ];
+  const [isAdmin, setIsAdmin] = useState(false);
+  const [isActive, setIsActive] = useState(false);
 
-  const colorOptions = [
-    { label: "Red", value: "red" },
-    { label: "Blue", value: "blue" },
-    { label: "Green", value: "green" },
-    { label: "Yellow", value: "yellow" },
-    { label: "Purple", value: "purple" },
-    { label: "Orange", value: "orange" },
-  ];
+  const handleSubmit = (e: React.FormEvent<HTMLFormElement>) => {
+    e.preventDefault();
+
+    alert(JSON.stringify({
+      isAdmin,
+      isActive,
+    }, null, 2));
+  };
+
 
   return (
-    <Form onSubmit={onSubmit}>
-      {({ formState: { isSubmitting } }) => (
-        <div className="flex w-[400px] flex-col gap-4">
-          <EnumField
-            name="gender"
-            label="Gender"
-            variant="RadioGroup"
-            options={genderOptions}
-            required
-          />
+    <div className="flex gap-4">
+      <Form onSubmit={onSubmit} className="bg-gray-100 rounded-md">
+        {({ formState: { isSubmitting } }) => (
+          <div className="flex w-[400px] flex-col gap-4 p-2">
+            <BooleanField name="isAdmin" label="Is Admin" />
+            <BooleanField name="isActive" label="Is Active" isToggle />
 
-          <EnumField
-            name="favoriteColor"
-            label="Favorite Color"
-            description="Choose your favorite color"
-            variant="Select"
-            options={colorOptions}
-            searchable
-            placeholder="Select a color"
-            required
-          />
+            <Button type="submit" disabled={isSubmitting}>
+              {isSubmitting ? "Submitting..." : "Submit"}
+            </Button>
+          </div>
+        )}
+      </Form>
+      <div className="flex flex-col gap-4 w-full">
+        <div className="bg-gray-100 rounded-md p-2">
+          <form onSubmit={handleSubmit} className="flex w-[400px] flex-col gap-4">
+            <Checkbox name="isAdmin" label="Is Admin" value={isAdmin} onChange={() => setIsAdmin(!isAdmin)} />
+            <Toggle name="isActive" label="Is Active" value={isActive} onChange={() => setIsActive(!isActive)} />
 
-          <Button type="submit" disabled={isSubmitting}>
-            {isSubmitting ? "Submitting..." : "Submit"}
-          </Button>
+            <Button type="submit">
+              Submit
+            </Button>
+          </form>
         </div>
-      )}
-    </Form>
+      </div>
+    </div>
   );
 };
